Refactor pyqt5.py in to a series of smaller files
All view related files will now be placed in: src/view/<framework>
