+++ src/view/pyqt5/main.py
@@ -1,0 +1,201 @@
+from PyQt5.QtWidgets import QMainWindow, QWidget, QVBoxLayout, QHBoxLayout, QLabel, QListWidget, QAbstractItemView, QAction, QListWidgetItem
+from PyQt5.QtGui import QIcon, QFont, QPixmap
+from PyQt5.QtCore import Qt
+from PyQt5.QtSvg import QSvgWidget
+from src import constants
+from src.executer_arduino import ArduinoPassthroughExecuter
+from src.view.pyqt5.filter_dialog import FilteredListDialog
+from src.view.pyqt5.edit_loadout_dialog import EditLoadoutDialog
+
+class MainWindow(QMainWindow):
+    def __init__(self, controller):
+        super().__init__()
+        self.controller = controller
+        self.setWindowTitle("Hell snake")
+        self.setWindowIcon(QIcon(constants.ICON_BASE_PATH+"hell_snake.png"))
+        self.setMinimumSize(350, 225)
+        self.resize(350,225)
+        self.setWindowFlags(Qt.WindowStaysOnTopHint)
+
+        central_widget = QWidget()
+        central_widget.setStyleSheet("background-color: white")
+        self.vBox = QVBoxLayout()
+        self.vBox.setSpacing(0)
+        self.vBox.setContentsMargins(0,0,0,0)
+        central_widget.setLayout(self.vBox)
+        self.setCentralWidget(central_widget)
+
+        self.hBox = QHBoxLayout()
+        self.hBox.setContentsMargins(0,5,0,0)
+        self.armedIcon = QLabel()
+        self.armedIcon.setFixedSize(70, 70)
+        self.armedIcon.setMargin(5)
+        self.armedIcon.setAlignment(Qt.AlignCenter)
+        self.armedIcon.setScaledContents(True)
+        self.hBox.addWidget(self.armedIcon)
+
+        self.loadout = QLabel()
+        self.loadout.setFixedHeight(70)
+        self.loadout.setAlignment(Qt.AlignVCenter|Qt.AlignLeft)
+        font = QFont("Arial", 24)
+        font.setBold(True)
+        self.loadout.setFont(font)
+        self.hBox.addWidget(self.loadout)
+        self.vBox.addLayout(self.hBox)
+
+        self.armedBar = QLabel()
+        self.armedBar.setFixedHeight(10)
+        self.vBox.addWidget(self.armedBar)
+
+        self.listwidget = QListWidget()
+        self.listwidget.setSelectionMode(QAbstractItemView.NoSelection)
+        self.listwidget.setFocusPolicy(Qt.NoFocus)
+        self.listwidget.itemClicked.connect(self.on_macro_clicked)
+        self.vBox.addWidget(self.listwidget)
+
+        self.setup_toolbar_menu()
+        self.update_armed()
+    
+    def on_strategem_selected(self, key, id):
+        self.controller.update_macro_binding(key, id)
+
+    def on_macro_clicked(self, item):
+        dialog = FilteredListDialog(self.controller, item.data(Qt.UserRole), self.on_strategem_selected)
+        dialog.exec_()
+
+    def update_macros(self):
+        self.listwidget.clear()
+
+        for index, (key, value) in enumerate(self.controller.model.macros.items()):
+            listAdapter = QLoadoutListAdapter()
+            listAdapter.setKey(key)
+            listAdapter.setStyleSheet("background-color: transparent")
+            listAdapter.setStrategem(value)
+
+            listAdapterItem = QListWidgetItem(self.listwidget)
+            listAdapterItem.setData(Qt.UserRole,key)
+            listAdapterItem.setSizeHint(listAdapter.sizeHint())
+            self.listwidget.addItem(listAdapterItem)
+            self.listwidget.setItemWidget(listAdapterItem, listAdapter)
+
+        height = 115 + (len(self.controller.model.macros)*55)
+        self.resize(self.geometry().width(), height)
+
+    def update_armed(self):
+        if self.controller.model.isArmed:
+            self.armedIcon.setPixmap(QPixmap(constants.ICON_BASE_PATH+"armed.png"))
+            self.armedBar.setStyleSheet("background-color: red")
+        else: 
+            self.armedIcon.setPixmap(QPixmap(constants.ICON_BASE_PATH+"disarmed.png"))     
+            self.armedBar.setStyleSheet("background-color: gray")
+    
+    def update_current_loadout(self):
+        currentLoadout = self.controller.model.currentLoadout
+        if currentLoadout is not None:
+            self.loadout.setText(currentLoadout.name) 
+        else:
+            self.loadout.setText("")
+
+    def setup_toolbar_menu(self):
+        # Create a Files menu
+        files_menu = self.menuBar().addMenu("Files")
+
+        settingsOptions = files_menu.addMenu(QIcon(constants.ICON_BASE_PATH+"settings.svg"), "Settings")
+
+        print_action = QAction(QIcon(constants.ICON_BASE_PATH+"settings_print.svg"), "Print settings", self)
+        print_action.triggered.connect(self.controller.print_settings)
+        settingsOptions.addAction(print_action)
+
+        save_action = QAction(QIcon(constants.ICON_BASE_PATH+"settings_save.svg"), "Save settings", self)
+        save_action.triggered.connect(self.controller.save_settings)
+        settingsOptions.addAction(save_action)
+
+        files_menu.addSeparator()
+
+        exit_action = QAction(QIcon(constants.ICON_BASE_PATH+"exit.svg"),"Exit", self)
+        exit_action.triggered.connect(self.controller.exit)
+        files_menu.addAction(exit_action)
+
+        arm_action = QAction("Arm", self)
+        arm_action.triggered.connect(self.controller.toggle_armed)
+        self.menuBar().addAction(arm_action)
+
+        self.loadout_menu = self.menuBar().addMenu("Loadouts")
+        self.update_loadout_menu_items()
+    
+    def update_loadout_menu_items(self):
+        self.loadout_menu.clear()
+        for loadoutId, loadout in self.controller.model.settings.loadouts.items():
+            loadout_action = QAction(loadout.name, self)
+            loadout_action.triggered.connect(lambda checked, loadoutId=loadoutId: self.controller.set_active_loadout(loadoutId))
+            self.loadout_menu.addAction(loadout_action)
+        
+        self.loadout_menu.addSeparator()
+
+        loadout_edit_action = QAction(QIcon(constants.ICON_BASE_PATH+"edit_loadout.svg"), "Edit loadouts", self)
+        loadout_edit_action.triggered.connect(self.open_edit_loadout_dialog)
+        self.loadout_menu.addAction(loadout_edit_action)
+
+    def add_executor_settings(self):
+        executor = self.controller.executer
+        if isinstance(executor,ArduinoPassthroughExecuter):
+            self.select_serial = self.menuBar().addMenu("Select serial")
+            self.update_serial_menu()
+    
+    def update_serial_menu(self):
+        executor = self.controller.executer
+        self.select_serial.clear()
+        connection = executor.get_current_connection()
+        physical_addresses = executor.get_physical_addresses()
+        for port, desc, hwid in sorted(physical_addresses):
+            serial = QAction(desc, self)
+            serial.triggered.connect(lambda checked, port=port: self.connect_to_serial(port))
+            if port == connection:
+                serial.setIcon(QIcon(constants.ICON_BASE_PATH+"serial_connected")) # TODO Change icon to a selected icon
+            self.select_serial.addAction(serial)
+    
+    def connect_to_serial(self, port):
+        self.controller.executer.connect_to_arduino(port)
+        self.update_serial_menu()
+
+    def open_edit_loadout_dialog(self):
+        dialog = EditLoadoutDialog(self.controller)
+        dialog.exec_()
+
+class QLoadoutListAdapter(QWidget):
+    def __init__ (self, parent = None):
+        super(QLoadoutListAdapter, self).__init__(parent)
+
+        self.hBox  = QHBoxLayout()
+        self.hBox.setContentsMargins(5,5,5,0)
+        
+        self.icon = QLabel()
+        self.icon.setFixedSize(50, 50)
+        self.icon.setAlignment(Qt.AlignCenter)
+        self.icon.setStyleSheet("background-color: #ff1f2832")
+        self.hBox.addWidget(self.icon)
+        
+        self.key = QLabel()
+        self.key.setFixedSize(50, 50)
+        self.key.setAlignment(Qt.AlignCenter)
+        font = QFont("Arial", 24)
+        font.setBold(True)
+        self.key.setFont(font)
+        self.hBox.addWidget(self.key)
+        
+        self.name = QLabel()
+        self.name.setFixedHeight(50)
+        self.hBox.addWidget(self.name)
+           
+        self.setLayout(self.hBox)
+
+    def setStrategem(self, strategem):
+        self.name.setText(strategem.name)
+        svg_widget = QSvgWidget(constants.STRATEGEM_ICON_PATH+strategem.icon_name)
+        svg_widget.setFixedSize(40,40)
+        svg_widget.setStyleSheet("background-color: transparent")
+        self.icon.setPixmap(svg_widget.grab())  
+    
+    def setKey(self, key):
+        self.id = key
+        self.key.setText(key)
