+++ src/view/view_base.py
